backend/mysite/requirements.txt
Django==3.2
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.5.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.21.1
reportlab==3.6.12
requests==2.28.2